contrib/rootfs-builder: Support xz (and other) compression formats
I'm not sure when the change happened (sometime since ff5e5788,
contrib/rootfs-builder: Support multiple architectures, 2017-09-20,
#479), but Gentoo is now using xz compression for some amd64 stages:

  $ curl http://distfiles.gentoo.org/releases/amd64/autobuilds/latest-stage3.txt
  # Latest as of Fri, 09 Mar 2018 15:00:02 +0000
  # ts=1520607602
  20180308T214502Z/stage3-amd64-20180308T214502Z.tar.xz 188851072
  20180308T214502Z/hardened/stage3-amd64-hardened-20180308T214502Z.tar.xz 174759140
  20180308T214502Z/hardened/stage3-amd64-hardened+nomultilib-20180308T214502Z.tar.xz 167639136
  20180301T214503Z/hardened/stage3-amd64-hardened-selinux-20180301T214503Z.tar.xz 187319664
  20180308T214502Z/stage3-amd64-nomultilib-20180308T214502Z.tar.xz 181866296
  20180303/systemd/stage3-amd64-systemd-20180303.tar.bz2 281873162
  20180225/uclibc/stage3-amd64-uclibc-hardened-20180225.tar.bz2 57147392
  20180225/uclibc/stage3-amd64-uclibc-vanilla-20180225.tar.bz2 144490332
  20180308T214502Z/stage3-x32-20180308T214502Z.tar.xz 205285160
  20180308T214502Z/hardened/stage4-amd64-hardened+minimal-20180308T214502Z.tar.bz2 261776436
  20180308T214502Z/hardened/stage4-amd64-hardened+minimal-nomultilib-20180308T214502Z.tar.bz2 251821245
  20180308T214502Z/stage4-amd64-minimal-20180308T214502Z.tar.bz2 280918565
  20180308T214502Z/stage4-amd64-minimal-nomultilib-20180308T214502Z.tar.bz2 268835860
  20180303/systemd/stage4-amd64-systemd-20180303.tar.bz2 622067712

This commit adapts to that change, allowing us to continue to pull the
vanilla stage3-amd64-20180308T214502Z.tar.xz.

+++ contrib/rootfs-builder/get-stage3.sh
@@ -78,13 +78,23 @@ then
 			die "could not calculate DATE from ${BASE_ARCH_URL}latest-stage3.txt"
 		fi
 	fi
-	STAGE3="${STAGE3:-stage3-${STAGE3_ARCH}-${DATE}.tar.bz2}"
+
+	STAGE3=$(echo "${LATEST}" | sed -n "s|${DATE}/\(stage3-${STAGE3_ARCH}-${DATE}[.]tar[.][^ ]*\) .*|\1|p")
+	if test -z "${STAGE3}"
+	then
+		die "could not calculate STAGE3 from ${BASE_ARCH_URL}latest-stage3.txt"
+	fi
 fi
 
 ARCH_URL="${ARCH_URL:-${BASE_ARCH_URL}${DATE}/}"
 STAGE3_CONTENTS="${STAGE3_CONTENTS:-${STAGE3}.CONTENTS}"
 STAGE3_DIGESTS="${STAGE3_DIGESTS:-${STAGE3}.DIGESTS.asc}"
 
+COMPRESSION=$(echo "${STAGE3}" | sed -n 's/^.*[.]\([^.]*\)$/\1/p')
+if test -z "${COMPRESSION}"
+then
+	die "could not calculate COMPRESSION from ${STAGE3}"
+fi
 for FILE in "${STAGE3}" "${STAGE3_CONTENTS}" "${STAGE3_DIGESTS}"; do
 	if [ ! -f "downloads/${FILE}" ]; then
 		wget -O "downloads/${FILE}" "${ARCH_URL}${FILE}"
@@ -94,7 +104,11 @@ for FILE in "${STAGE3}" "${STAGE3_CONTENTS}" "${STAGE3_DIGESTS}"; do
 		fi
 	fi
 
-	CURRENT=$(echo "${FILE}" | sed "s/${DATE}/current/")
+	FILE_NOCOMPRESSION=$(echo "${FILE}" | sed "s/[.]${COMPRESSION}//")
+	if [ "${FILE_NOCOMPRESSION}" = "${FILE}" ]; then
+		die "unable to remove .${COMPRESSION} from ${FILE}"
+	fi
+	CURRENT=$(echo "${FILE_NOCOMPRESSION}" | sed "s/${DATE}/current/")
 	(
 		cd downloads &&
 		rm -f "${CURRENT}" &&
@@ -102,4 +116,3 @@ for FILE in "${STAGE3}" "${STAGE3_CONTENTS}" "${STAGE3_DIGESTS}"; do
 		die "failed to link ${CURRENT} -> ${FILE}"
 	)
 done
-
